Refacto bank example
