+++ event_store/src/storage/postgres.rs
@@ -26,8 +26,9 @@ impl PostgresBackend {
     /// # Errors
     ///
     /// In case of Postgres connection error
+    #[tracing::instrument(name = "PostgresBackend", skip(url))]
     pub async fn with_url(url: &str) -> Result<Self, sqlx::Error> {
-        trace!("Connecting a new PostgresBackend");
+        trace!("Creating");
 
         Ok(Self {
             pool: PgPoolOptions::new()
